StableDiffusionWrapper/requirements.txt
 fastapi
 uvicorn
 torch
 transformers
 diffusers
 accelerate
 sentencepiece
 protobuf

 python-dotenv